Simplify custom section style configuration
Use a shorter name for the model that should be configured. Instead of having to
type out the entire class name (e.g. "KnowledgeBase::Sectionables::Text"), you
would now drop the "KnowledgeBase::Sectionables" part and snake_case the name as
well as using a symbol, which we've all learned to love so much.